fixed empty serializator finalization
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@7144 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ mrt/serializator.cpp
@@ -267,6 +267,10 @@ const Chunk & Serializator::getData() const {
 }
 
 void Serializator::finalize(mrt::Chunk &data) {
+	if (_data->empty()) {
+		data.free();
+		return;
+	}
 	data.setData(_data->getPtr(), _data->getSize(), true);
 	_data->unlink();
 }
